requirements_qalign.txt
accelerate==0.25.0
asttokens==3.0.1
certifi==2022.12.7
charset-normalizer==2.1.1
colorama==0.4.6
contourpy==1.3.2
cycler==0.12.1
einops==0.8.1
executing==2.2.1
filelock==3.20.0
fonttools==4.61.1
fsspec==2025.12.0
hf-xet==1.2.0
huggingface-hub==0.36.0
icecream==2.1.8
idna==3.4
Jinja2==3.1.6
kiwisolver==1.4.9
MarkupSafe==2.1.5
matplotlib==3.10.8
mpmath==1.3.0
networkx==3.4.2
ninja==1.13.0
numpy==1.26.4
opencv-python==4.9.0.80
packaging==25.0
pandas==2.3.3
pillow==12.0.0
protobuf==6.33.2
psutil==7.1.3
Pygments==2.19.2
pyparsing==3.3.1
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.3
regex==2025.11.3
requests==2.28.1
safetensors==0.7.0
scipy==1.15.3
sentencepiece==0.2.1
six==1.17.0
sympy==1.14.0
tokenizers==0.15.2
torch==2.1.2+cu118
torchaudio==2.1.2+cu118
torchvision==0.16.2+cu118
tqdm==4.67.1
transformers==4.36.2
triton==2.1.0
typing_extensions==4.15.0
tzdata==2025.3
urllib3==1.26.13
